+++ loess-rust/src/names.rs
@@ -1,0 +1,2 @@
+mod visibility_and_privacy;
+pub use visibility_and_privacy::*;